src/requirements.txt
asgiref==3.9.1
certifi==2025.7.14
charset-normalizer==3.4.2
dj-database-url==3.0.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
filelock==3.18.0
fsspec==2025.7.0
gunicorn==23.0.0
idna==3.10
jinja2==3.1.6
lxml==5.4.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.1
packaging==25.0
Pillow==11.3.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-docx==1.1.2
python-dotenv==1.0.1
requests==2.32.4
sqlparse==0.5.3
sympy==1.14.0
torch==2.7.1
torchvision==0.22.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.9.0
